# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "API endpoints for managing hotel bookings",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all bookings",
                "operationId": "afb8b225fc78962125d01a6c8e8af2ae",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingWithId"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "2911657b7059f988a4bc398a1d21e68f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingPOST"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get a booking by ID",
                "operationId": "4b13c909d897a5edd245741a4aafd9d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingWithId"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update a booking by ID",
                "operationId": "6f58c5d17bbc7034b229d97390054811",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingPUT"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a booking by ID",
                "operationId": "dc3a5e70d8c1811d5888122376cf81b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking deleted successfully"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookingPOST": {
                "title": "Booking for POST",
                "description": "Booking object",
                "required": [
                    "room_id",
                    "customer_id",
                    "check_in_date",
                    "check_out_date"
                ],
                "properties": {
                    "room_id": {
                        "description": "ID of the room booked",
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "description": "ID of the customer who made the booking",
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_date": {
                        "description": "Check-in date for the booking",
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-13"
                    },
                    "check_out_date": {
                        "description": "Check-out date for the booking",
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-15"
                    }
                },
                "type": "object"
            },
            "BookingPUT": {
                "title": "Booking for PUT",
                "description": "Booking object",
                "properties": {
                    "room_id": {
                        "description": "ID of the room booked",
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "description": "ID of the customer who made the booking",
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_date": {
                        "description": "Check-in date for the booking",
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-13"
                    },
                    "check_out_date": {
                        "description": "Check-out date for the booking",
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-15"
                    }
                },
                "type": "object"
            },
            "BookingWithId": {
                "title": "Booking with ID",
                "description": "Booking object",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the booking",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "room_id": {
                        "description": "ID of the room booked",
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "description": "ID of the customer who made the booking",
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_date": {
                        "description": "Check-in date for the booking",
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-13"
                    },
                    "check_out_date": {
                        "description": "Check-out date for the booking",
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-15"
                    },
                    "created_at": {
                        "description": "Timestamp when the booking was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-13 12:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp when the booking was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-13 12:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    }
}